shader_runner/spirv: add check for SPV extension support
This only affects the case where SPIR-V shaders are run, specially
when using tests coming from other extensions.

In several cases, OpenGL supports an extension with a given
functionality. And such functionality it is not supported on core
SPIR-V but on a equivalent SPV functionality.

But it is not mandatory to support such SPV extension even if you
support the equivalent OpenGL one.

So for that case, we need to check for the equivalent SPV extension
and check if it is supported.

Acked-by: Marek Olšák <marek.olsak@amd.com>
Acked-by: Timothy Arceri <tarceri@itsqueeze.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1>