generated_tests: Add script to translate shader_tests to SPIR-V
Utility script which translates GLSL shader_tests to OpenGL SPIR-V
shader_tests (ARB_gl_spirv), using glslang and SPIR-V tools.

The translation to SPIR-V is done using glslangValidator with the
--auto-map-locations option.

Names for uniforms, UBOs and vertex attributes in the [test] and
[vertex data] sections are remapped. In order to do that, the
generated SPIR-V is parsed to get the locations, bindings, offsets,
matrix stride and row major, array indices and the tree of types.

--aml can’t handle allocating the same varying locations across
multiple stages because it only sees the shaders for one stage at a
time. To work around this the script reads overwrites the locations
for the inputs in the generated SPIR-V with the locations with the
same name in the outputs from the previous stage.

Based on Nicolai Hänle's work who wrote the script when glslang did
not provide the --auto-map-location option and the GLSL code was
parsed by the script. The script was modified to use such option,
remove the GLSL parsing and implement the remapping of uniforms, UBOs
and vertex attributes, add more command line options and other
improvements.

Signed-off-by: Nicolai Hänle <nicolai.haehnle@amd.com>
Signed-off-by: Alejandro Piñeiro <apinheiro@igalia.com>
Signed-off-by: Neil Roberts <nroberts@igalia.com>
Signed-off-by: Antia Puentes <apuentes@igalia.com>

Acked-by: Marek Olšák <marek.olsak@amd.com>
Acked-by: Timothy Arceri <tarceri@itsqueeze.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1>