arb_gl_spirv: Add ARB_program_interface_query tests
Added tests to check that the ARB_program_interface_query queries
involving names behave as described in the ARB_gl_spirv extension.

The SPIR-V used do not contain names. As names are optional in SPIR-V
this is useful to test the behaviour of the name related queries when
no name reflection is available.

Adds also tests to check that the queries not involving names work as
normally, as defined in the ARB_gl_spirv_extension specification.

Acked-by: Marek Olšák <marek.olsak@amd.com>
Acked-by: Timothy Arceri <tarceri@itsqueeze.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1>