arb_gl_spirv: Add test equivalent to CS basic-ssbo
The difference with
arb_compute_shader/execution/basic-ssbo.shader_test is that we
explicitly assign a binding to the SSBO.  Without a binding number
specified, glslang implicitly assigns a binding number of 1 to the
SSBO, while Mesa assigns a binding number of 0. To ensure the test
runs consistently in both cases, assign the binding explicitly.

Acked-by: Marek Olšák <marek.olsak@amd.com>
Acked-by: Timothy Arceri <tarceri@itsqueeze.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1>