arb_gl_spirv: add a xfb test that uses different buffers, mixing order
In order to check that GL_TRANSFORM_FEEDBACK_BUFFER_INDEX is properly
computed, as at some point, when mixing order it was assigned extra
buffer_index.

It is worth to note that although the original GLSL do that definition
out, glslang groups and reorder the definitions based on xfb_offset
and xfb_buffer, so we needed to tweak by hand the SPIR-V generated.

Acked-by: Marek Olšák <marek.olsak@amd.com>
Acked-by: Timothy Arceri <tarceri@itsqueeze.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1>