arb_occlusion_query: execute test even if no query bits
It is legal for a driver to support the query API but not have
any query bits. We used to just skip the test in that case.

However, drivers must still allow calling the API entrypoints, so
let's test that but ignore the result.

While there's more tests that early-outs like this, it's probably
sufficient that a single test checks this.

Reviewed-by: Emma Anholt <emma@anholt.net>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/746>