line-smooth-stipple: do not test undefined behavior
The initial value of the stipple-counter for a clipped line is
indeterminate, according to the OpenGL spec. So let's make sure the
line doesn't need clipping, by offsetting in the oposite direction, and
adding a one-pixel margin on the left and right sides.

This fixes the test on Softpipe, which only updates the stipple-counter
for visible fragments.

Reviewed-by: Roland Scheidegger <sroland@vmware.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/760>