perf: port glsl-compile-time perf-test from mesa demos
This test wasn't using the common perf-helpers, and I'm not really sure
it makes sense either, because it's testing timings of side-effects like
fixed-function shader compiles that won't repeat if the code to render
is repeated.

So let's just keep this test using open-coded time-logging instead.
Luckily, our nano-second timers are plenty accurate to capture the time
this takes on computers for the forseeable future.

Reviewed-by: Adam Jackson <ajax@redhat.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/764>