shaders@glsl-max-varyings: reduce size of pixels
With a pixel size of 2x2 size the drawing area is 1026x1026, added
to that window decorations, and possible offsets from the desktop
environment the window might not fit on the screen so that the desktop
manager might scale the window to fit the screen. Running the piglit
without the -fbo flag will then lead to reading back the output from
the wrong location.

Reduce the size of the pixels to 1x1 and drop the padding so that we end
up with a drawing area of 256x256, which is more likely to fit on the screen.

v2: Drop the padding (Emma)
v3: use one glReadPixels (Emma)

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/771>