glsl-1.30: change offset for gl_ClipDistance test
With the offsets used to test gl_ClipDistance for the GL_LINES
primitives, the test fails for Raspberry Pi 4 device because the
rendered line is shifted 1 pixel to the left.

This is caused because the device uses 6 bits for subpixel precision in
the line rasterizer, so the final line is displaced 1 pixel regarding
test expects to be rendered.

Slighlty changing the offset makes the line to be rendered where we
want, and it doesn't affect the purpose of the test, which is verifying
that gl_ClilpDistance affects the primitive types correctly.

Reviewed-by: Emma Anholt <emma@anholt.net>
Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/806>