texturing: add a test for GPU/CPU sync when creating textures
On Panfrost, at least, a race condition could occur if a texture
is cleared and then initialized with texSubImage; the clear on
the GPU could happen after the memcpy in texSubImage. This test
checks for that and triggers the original Panfrost bug.

Signed-off-by: Eric R. Smith <eric.smith@collabora.com>
Reviewed-by: Erik Faye-Lund <erik.faye-lund@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/901>