clear-stencil: new test to check for problems clearing stencils
On some architectures clearing depth and stencils can be optimized, but
tracking the state to know when the stencil has been written is necessary
and sometimes has bugs. This test finds a problem with state tracking of
stencil clears in an older version of the panfrost driver.

Signed-off-by: Eric R. Smith <eric.smith@collabora.com>
Reviewed-by: Erik Faye-Lund <erik.faye-lund@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/905>