arb_texture_view: test for NaN in more cases
When the image has a float format, the test shader will be using
float texelFetch to test it. When we're clearing the image with a
NaN bit pattern, this becomes a problem as some hardware like ARM
Mali does not preserve NaN sign bit (or possibly the entire bit
pattern) so a texel that had 0x7fffffff written into it comes up
as 0xffffffff.

Make the test take a test_nan path in cases when NaN bit pattern
will be written into a float-typed image as NaN propagation
details should not affect the outcome of the test.

Acked-by: Marek Olšák <marek.olsak@amd.com>
Reviewed-by: Erik Faye-Lund <erik.faye-lund@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/912>