OVR_multiview: Add fb_shader_mismatch_draw_errors test
Add test that mismatches between multiview framebuffers and the
num_views declared in shaders result in the appropriate errors after
draw commands.

The test takes 2 arguments, the number of framebuffer views and the
number of views declared in the vertex shader. A glClear, triangle draw,
and glFinish are performed, and the appropriate error is checked for.

The test is run with a variety of arguments between 0 (non-multiview)
and 3 (3 views).

Tests between non-multiview and single multiview are omitted due to
uncertainty or abiguity over whether a non multiview FB/shader counts as
having 1 view or none?

Signed-off-by: James Hogan <james@albanarts.com>
Acked-by: Marek Olšák <marek.olsak@amd.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/987>