arb_program_interface_query: set vs_input2[1][0] as valid name
ARB_program_interface_query, question 8:

For GetProgramResourceIndex, what sort of strings can be used to match
the resource names for arrays of arrays?  For example, let's say a
shader declares a uniform such as:

uniform vec4 a[3][4][5];

Which one of the following names are accepted?  "a[2][1][0]" to
identify the base of the bottom-level array?  "a[2][1]", to identify
the same without the final "[0]"?  "a[2]", equivalent to "a[2][0][0]"?
Just "a", equivalent to "a[0][0][0]"?

RESOLVED:  We only accept entries of the form "a[2][1][0]" or
"a[2][1]", which is consistent with the existing rules that only allow
applications to omit the last index of a bottom-level array that has
been rolled up.

According to the resolution above, the possible names for
vec4 vs_input2[2][2]; are:

vs_input2[0]
vs_input2[0][0]
vs_input2[1]
vs_input2[1][0]

Signed-off-by: Mykhailo Skorokhodov <mykhailo.skorokhodov@globallogic.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/988>