arb_direct_state_access: disable dithering
Dithering is enabled by default in OpenGL but the test checks for
results assuming that no dithering is used.

Fixes gettextureimage-formats with format RGB5A1 on Panfrost when
the init-by-rendering parameter is used to initialize the texture by
rendering:

Panfrost uses an internal format with higher bit precision than
RGB5A1 to represent the framebuffer for the texture. Particularly, it
has more than 1 alpha bit and it will take advantage of that to apply
dithering on the writeback when dithering is enabled, causing the alpha
on the texture to have a dithered pattern and breaking the test
expectation.

Reviewed-by: Alejandro Piñeiro <apinheiro@igalia.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/990>