Don't rely on old numpy promotion rules in packing tests
The numeric promotion rules changed in NumPy 2.0.  We were previously
coding the packing functions as something like:

    uint32(y << 16 | x)    where x and y had type np.uint16

Apparently that used to work, where y << 16 would get promoted to an
np.uint32 and the final result would be an np.uint32.  With NumPy 2.x,
that is no longer the case, and the y << 16 | x calculation is done as
np.uint16's, so y's bits were totally lost.

The fact that it used to work at all is rather surprising.

We now promote each value to be packed to a full uint32 before or'ing.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/995>