replay: Adapt to new S3 proxy
The S3 proxy installed as part of the new infrastructure has some
exciting properties. The first one is that it's not actually S3, but
that's not relevant here.

The core issue is that it, after applying policies, returns a redirect
to the underlying (actually S3 API) storage service with a short-lived
authentication token in the URI. The tokens are the only method of
authentication, not the Authorization header.

Normally on cross-host redirects, HTTP clients will remove an
Authorization token to avoid leaking credentials across domains.
However, when we're using an S3 proxy, we will redirect from
http://caching-proxy/cache/?uri=s3.fd.o/mesa-rootfs/foo to
http://caching-proxy/cache/?uri=hel1.your-storageservice.net/fdo-opa/mesa-rootfs/foo.
As this does not cross a domain boundary, python-requests will not strip
the Authorization header, which Hetzner responds to with a 400.

Get around this by implementing our own HTTP redirect handling,
stripping the Authorization header whilst we chase the redirects.

Signed-off-by: Daniel Stone <daniels@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/996>