khr_texture_compression_astc: normalize error-color
The spec says that when using linear RGB decode, the error color can
either be magenta or all-NaN. So let's make sure we allow either.

This is slightly more lenient than needed, because we use the same
shader for the sRGB and linear RGB tests. But the complexity of
duplicating the shaders etc just doesn't seem to be worth it, so
let's just be tad lax here.

Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1012>