replay: Calculate the chunk size via etag suffix
On S3, a '-' in the ETag header indicates a multipart upload, with the
number on the right representing the number of chunks used for the ETag
calculation. Although we usually use a 5 MB chunk size, this may differ
depending on the upload process.

Let's calculate the chunk size using the metadata when available,
falling back to the default size if not specified.

Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1030>