framework: add and use TestPlaceholder objects
The actual Test classes are not that big, but with deqp-vk creating
around 2800 instances currently, we quickly consume around 3GB of ram
before beginning any tests. Given that a lot of these objects may not
even be used, we can reduce memory overhead by using a placeholder
namedtuple that is expanded to a full Test instance when requested from
the TestDict class. This cuts the initial memory usage back to 1.6 GB.

This does add a small lookup penalty, but given the lookup penalty is
still smaller that the time it takes to create an instance of a Test, it
results in a net improvement for most subsets of deqp tests. Not to
mention this penalty can be split between threads, unlike previously
when it was occurring in the single-threadded profile creation step.

The following data was collected using a subset of 3151 deqp-vk tests:

Threads   Test Execution Time (s)   TestPlaceholder Execution Time (s)
-------   -----------------------   ----------------------------------
1         2613.93                   2599.95
8          413.31                    400.01
32         299.75                    286.73

Reviewed-by: Christian Gmeiner <cgmeiner@igalia.com>
Signed-off-by: Randolph Sapp <rs@ti.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1032>